+++ sheerwater_benchmarking/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utility functions for benchmarking."""
-from .cachable import cacheable
+from .caching import cacheable
 from .remote import dask_remote
 from .secrets import cdsapi_secret, ecmwf_secret, salient_secret, salient_auth
 from .data_utils import apply_mask, roll_and_agg, regrid
